fix application-conf lazy loading fix.
