+++ build.sbt
@@ -1,20 +1,28 @@
 import CompilerSettings._
+import com.typesafe.sbt.GitVersioning
+import TodoListPlugin._
 
-lazy val projectSettings = Seq(
-    organization := "coop.rchain",
-    name := "rsong-acquisition",
-    scalaVersion := "2.12.6"
-  )
+
+val gitCommitString = SettingKey[String]("gitCommit")
+
+gitCommitString := git.gitHeadCommit.value.getOrElse("Not Set")
 
 lazy val root = (project in file("."))
-  .settings(projectSettings: _*)
-  .settings(
-    libraryDependencies ++= {
+  . settings(
+    buildInfoKeys := Seq[BuildInfoKey](version, gitCommitString),
+    buildInfoPackage := "buildInfo",
+    buildInfoOptions += BuildInfoOption.ToMap,
+    buildInfoOptions += BuildInfoOption.ToJson
+  ).settings(
+  organization := "coop.rchain",
+  name := "rsong-acquisition",
+  scalaVersion := "2.12.6",
+  libraryDependencies ++= {
       object V {
         val specs2 = "4.2.0"
         val logback = "1.2.3"
         val scalalogging = "3.9.0"
-        val config = "1.3.3"
+        val config = "1.3.2"
         val scalapb= "0.7.4"
         val circie="0.9.3"
         val catsEffect="1.0.0"
@@ -41,7 +49,12 @@ scalacOptions in Compile ++= CompilerSettings.options
 PB.targets in Compile := Seq(
   scalapb.gen() -> (sourceManaged in Compile).value)
 
-enablePlugins(JavaServerAppPackaging)
+enablePlugins(JavaServerAppPackaging, BuildInfoPlugin)
+
+enablePlugins(GitVersioning)
+
+compileWithTodolistSettings
+
 
 dockerRepository := Some("kayvank")
 dockerUpdateLatest := true
