+++ project/plugins.sbt
@@ -1,7 +1,13 @@
+addSbtPlugin("org.lyranthe.sbt" % "partial-unification" % "1.1.0")
+addCompilerPlugin("org.spire-math" %% "kind-projector" % "0.9.6")
 addSbtPlugin("com.typesafe.sbt" % "sbt-native-packager" % "1.3.6")
 addSbtPlugin("com.thesamet" % "sbt-protoc" % "0.99.18")
 addSbtPlugin("com.geirsson" % "sbt-scalafmt"  % "1.4.0")
 addSbtPlugin("com.eed3si9n" % "sbt-buildinfo" % "0.7.0")
 addSbtPlugin("net.virtual-void" % "sbt-dependency-graph" % "0.9.0")
 addSbtPlugin("com.typesafe.sbt" % "sbt-site" % "1.3.2")
+addSbtPlugin("com.typesafe.sbt" % "sbt-ghpages" % "0.6.2")
 addSbtPlugin("se.marcuslonnberg" % "sbt-docker" % "1.5.0")
+addSbtPlugin( "com.eed3si9n" % "sbt-buildinfo" % "0.7.0")
+addSbtPlugin("com.github.fedragon" % "sbt-todolist" % "0.7")
+addSbtPlugin("com.typesafe.sbt" % "sbt-git" % "1.0.0")
