+++ src/main/resources/logback.xml
@@ -15,7 +15,7 @@
         </layout>
     </appender>
 
-    <root level="INFO">
+    <root level="DEBUG">
         <appender-ref ref="FILE" />
         <appender-ref ref="STDOUT" />
     </root>
