+++ src/main/resources/rsong-acquisition.conf
@@ -3,7 +3,7 @@ coop {
     rsong {
       assets.path:"/home/kayvan/dev/assets/RCHAIN Assets"
       assets.path: ${?RSONG_ASSETS_PATH}
-        contract.file.name: "/rho/rsong.rho"
+      contract.file.name: "/rho/rsong.rho"
        grpc {
          host: "localhost"
          host: ${?GRPC_SERVER}
@@ -17,7 +17,6 @@ coop {
          version: "v1"
          http {
             port: 9000
-            port: ${?HTTP_PORT}
          }
        }
     }
