--- src/main/scala/coop/rchain/utils/Globals.scala
@@ -1,16 +1,0 @@
-package coop.rchain.utils
-
-import com.typesafe.config.{Config, ConfigFactory}
-import coop.rchain.repo.RholangProxy
-
-object Globals {
-  lazy val cfg = ConfigFactory.load
-  val appCfg = cfg.getConfig("coop.rchain.rsong")
-
-  val artpath = "v1/art"
-  val songpath = "v1/song/music"
-//  val rsongHostUrl: String = "http://dev-rchain.com"
-  val rsongHostUrl: String = java.net.InetAddress.getLocalHost.getHostName
-  val rsongPath = "/home/kayvan/dev/assets/RCHAIN Assets"
-
-}
